+++ src/components/common/Button.tsx
@@ -1,98 +1,16 @@
-import { cva } from 'class-variance-authority'
 import type { ButtonHTMLAttributes } from 'react'
-import type { VariantProps } from 'class-variance-authority'
+
 import cn from '@/lib/cn'
 
-/**
- * 버튼 스타일 정의 (CVA)
- *
- * 기본 세팅값
- *
- * variant
- * - primary   : 보라색 버튼
- * - secondary : active 시 색상 변경
- * - disabled  : 비활성 전용 버튼
- * - kakao     : 카카오 로그인 버튼 (노란색)
- * - naver     : 네이버 로그인 버튼 (초록색)
- *
- * size
- * 사이즈에 맞는걸로 사용해주세요
- */
-export const buttonVariants = cva(
-  /**
-   * 공통 베이스 스타일
-   */
-  'inline-flex items-center justify-center border  transition-colors px-[20px] text-base',
-  {
-    variants: {
-      /**
-       * 버튼 색상 스타일
-       */
-      variant: {
-        primary:
-          'bg-primary text-white border-transparent hover:bg-primary-hover active:bg-primary-active',
+import { buttonVariants, type ButtonVariantProps } from './buttonVariants'
 
-        secondary:
-          'bg-primary-100 border border-primary text-primary hover:bg-primary-200 active:bg-primary-300',
-
-        disabled:
-          'bg-gray-200 text-gray-400 border-transparent cursor-not-allowed pointer-events-none',
-
-        kakao:
-          'bg-[#fee500] text-[#391C1A] border-transparent hover:bg-[#fdd835] active:bg-[#fbc02d] transition-colors',
-
-        naver:
-          'bg-[#03c75a] text-white border-transparent hover:bg-[#02b350] active:bg-[#029f47] transition-colors',
-      },
-
-      /**
-       * 버튼 사이즈
-       */
-      size: {
-        xxs: 'w-[78px] h-[42px]',
-        xs: 'w-[92px] h-[64px]',
-        sm: 'w-[112px] h-[48px]',
-        md: 'w-[126px] h-[48px]',
-        lg: 'w-[168px] h-[48px]',
-        xl: 'w-[348px] h-[52px]',
-        xxl: 'w-[480px] h-[52px]',
-      },
-      /**
-       * 버튼 라운드 값
-       */
-      rounded: {
-        default: 'rounded-[4px]',
-        full: 'rounded-full',
-      },
-    },
-
-    /**
-     * 기본값 설정
-     */
-    defaultVariants: {
-      variant: 'primary',
-      rounded: 'default',
-    },
-  }
-)
-export interface ButtonProps
-  extends ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {}
+export type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & ButtonVariantProps
 
 /**
  * 버튼 컴포넌트
- * 사용 예시
- * <Button
- *   variant="primary | ..."
- *   size="sm | md | ..."
- *   className={...}
- *   onClick={...}
- * >
- *   저장
- * </Button>
  *
- * disabled 스타일을 적용하려면
- * 반드시 variant="disabled"를 함께 사용해주세요.
+ * 사용 예: <Button variant="primary" size="sm">저장</Button>
+ * disabled 스타일은 variant="disabled"와 함께 사용.
  */
 export function Button({
   variant,
