+++ src/components/common/index.ts
@@ -9,3 +9,5 @@ export { default as Loading } from './Loading'
 export { Modal } from './Modal'
 export * from './Modal/variants'
 export { VerificationButton } from './VerificationButton'
+export { buttonVariants } from './buttonVariants'
+export type { ButtonVariantProps } from './buttonVariants'
